a program that finds and prints the first 98 Fibonacci numbers, starting with 1 and 2, followed by a new line.

+++ functions_nested_loops/104-fibonacci.c
@@ -10,7 +10,7 @@ int main(void)
 	unsigned long i;
 	unsigned long j;
 	unsigned long s;
-	int m, n;
+	unsigned long m, n;
 	int k;
 
 	i = 1;
